move functions from managers to nodes
