+++ src/lib/nodes/Endpoint.ts
@@ -1,3 +1,4 @@
+import ServiceMessage from '../messages/payloads/ServiceMessage';
 import { Node } from './Node';
 
-export class Endpoint extends Node {}
+export class Endpoint extends Node<ServiceMessage> {}
