+++ src/lib/nodes/managers/EndpointManager.ts
@@ -1,4 +1,3 @@
-import ServiceMessage from '../../messages/payloads/ServiceMessage';
 import { NodeManager } from './NodeManager';
 
-export class EndpointManager extends NodeManager<ServiceMessage> {}
+export class EndpointManager extends NodeManager {}
